Update Nikestore_ERD.md
Products are the shoes in the store. They are what the customers are there for.
Customers spend the money on your products, so you can order more.
Inventory is important because without enough product in the stores customers won't purchase anything.
Sale keeps track of customer and product info. 

Sales connect the customer and products, while inventory keeps products up with customers.
Keeping track of inventory levels helps the store not run out of shoes and miss sales. Keeping up with customer info will help you see a demand for certain products so you can order more for inventory.